Fixed scaling problems with Editable Transaction, rounding errors on JNumberField, and overflow errors on various areas.

+++ src/org/homeunix/drummer/view/logic/AccountListPanel.java
@@ -105,7 +105,7 @@ public class AccountListPanel extends ListPanelLayout {
 	}
 	
 	public AbstractBudgetPanel updateContent(){
-		int balance = 0;
+		long balance = 0;
 		
 		root.removeAllChildren(); 
 		selectedSource = null;
